# project-manager/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projects Manager API",
        "version": "0.1"
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "project"
                ],
                "description": "Displays a listing of projects",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "project"
                ],
                "description": "Stores a new project",
                "operationId": "store",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_name": {
                                        "description": "Name of the new project",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the new project",
                                        "type": "string"
                                    },
                                    "manager_id": {
                                        "description": "Id of the manager of the new project",
                                        "type": "integer"
                                    },
                                    "assigned_id": {
                                        "description": "Id of the assigned user of the new project",
                                        "type": "integer"
                                    },
                                    "enabled": {
                                        "description": "Status of the new project",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project stored successfully"
                    }
                }
            }
        },
        "/api/projects/{projectId}": {
            "get": {
                "tags": [
                    "project"
                ],
                "description": "Displays a specific project",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of project to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "project"
                ],
                "description": "Updates a specific project",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of project to edit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_name": {
                                        "description": "Updated name of the project",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated description of the project",
                                        "type": "string"
                                    },
                                    "manager_id": {
                                        "description": "Updated id of the manager of the project",
                                        "type": "integer"
                                    },
                                    "assigned_id": {
                                        "description": "Updated id of the assigned user of the project",
                                        "type": "integer"
                                    },
                                    "enabled": {
                                        "description": "Updated status of the project",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "project"
                ],
                "description": "Deletes a specific project",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of project to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}